자료구조 추가함.
